created cobra command for query, move old query function to internal/core and cmd/query now houses the cobra command only
